Form elements are now displayed
Update issue 79
Form elements are now displayed, with CSS generated content, a bug in the CSS parser prevented to do this before, but I solved it. Everything is still not perfect, but it is a good start. It will be in the next release or if you are impatient, it's in the SVN trunk.

git-svn-id: http://dompdf.googlecode.com/svn/trunk/dompdf@500 8e70de24-3a84-11de-b438-597f59cd7555