Handle globally-applied generated content (#758))
To sanely handle generated content applied globally using the
* wildcard we will only apply the content to teh body element
and its children.

When adding a new DOM node to an existing one check that the
existing node is represented in the frame tree. If not skip the
addition. This prevents an exception for certain nodes excluded
from the frame tree (e.g. HEAD content).