modify frame tree initialization loop to avoid skipping nodes
The recent iteration of the frame tree initialization code was prone to
skipping frames because the collection on which the loop was being run
was modified inside the loop. This version reuses code from an earlier
iteration and adds adjacent #text node concatenation to avoid unwanted
white space in the render.