Use basename/dirname instead of string-based path parsing
Parsing strings for path information was causing problems when
the directory separator didn't match the search character. This
character was a slash (/) in previous releases and the
DIRECTORY_SEPARATOR constant more recently. Rather than rely on
complicated character matching we'll use the built-in PHP
functions to parse the path information.

Fixes #1178