Return 0 if border radius is not set or empty
Previously we were checking for a value on the specified border
radius via `isset()`, but this condition is true for the default
value (an empty string). This resulted in the empty string being
run through `length_in_pt()`, which would then return returns the
default font size as the length.

Addresses an issue introduced by commit 1fa43aa