Add font fallback logic to the GD backend
Because we’re using the TTF GD methods we need to supply an actual font
file. However, since Dompdf relies on PDF viewer support for the core
fonts when rendering to PDF it does not include files for these fonts.
With this update the GD backend will substitute the relevant DejaVu
font when the specified font file can not be found.