Encode non-display text streams to UTF16
In order for CPDF to correctly support the full Unicode character set in metadata fields we need to encode to UTF16 with BOM. Previously the encoding would be based on the last-used font. Non-text streams in the PDF are not rendered using embedded fonts and so should be encoded without taking embedded font encoding into account.

This change also removes the font check from utf8toUtf16BE since this is a helper function and shouldn't need to check font support. That check is already performed by the caller.

Addresses #1514