Exclude nbsp from word boundary detection
When we added the Unicode flag to the text splitting regex it appears that the \s escape sequence was modified to include the full set of Unicode space separators. This set includes the no-break space (nbsp) character. With this update we're now using a double negative to select the same set of characters, minus the nbsp character.

Addresses #1759