Only use the style default if not set or inheritance has not been applied
Before this change we would use the default value if no computed value were present or if the specified property value was "inherit." However, if the computed value has been set we should use it. Except when its value is "inherit" in which case inheritance has not yet been applied.

This is a fallback, since in the normal Dompdf process flow inheritance will have been applied prior to accessing a style value via the getter.