Consider adjecent table cells border width when rendering background
Before this commit Dompdf would only consider the rendered borders of the cell being rendered. This was producing undesireable results when the adjacent cell had a larger border than the current cell, and when it was that cell's border being rendered. In that situation the current cell's background would render into the space occupied by the border.

This update considers the width of all borders rendered around the cell being rendered.

Addresses #1033