Trim white space after line new line text split
If we trim white space at a split before a line break has been applied then we may end up collapsing the space between words. And if we trim before determining the index at which to split we may end up with an infinite loop where the space character is put into a new frame in the same line over and over again. We only want to trim the white space from the current text frame if the split is going to result in a new line.

Also, we no longer need to add a space when calculating justified text line width. This was an artifact of the previous logic that did not always remove trailing spaces from a line.

Addresses #1174, #2188